serialize comparevalues helper
