+++ smart-contracts/test/LockSerializer/_compareValues.js
@@ -1,0 +1,30 @@
+const { ethers } = require('hardhat')
+
+const compareValues = async (serialized, lock) => {
+  const propNames = Object.keys(serialized)
+    .filter((k) => Number.isNaN(Number.parseInt(k))) // remove numbers from array index
+    .filter((k) => !['keyOwners', 'expirationTimestamps'].includes(k)) // exclude arrays
+  const values = await Promise.all(propNames.map((k) => lock[k]()))
+
+  // assertions
+  propNames.forEach((k, i) => {
+    if (
+      ethers.BigNumber.isBigNumber(serialized[k]) &&
+      ethers.BigNumber.isBigNumber(values[i])
+    ) {
+      assert.equal(
+        serialized[k].eq(values[i]),
+        true,
+        `different serialized value ${k}, ${serialized[k]}, ${values[i]}`
+      )
+    } else {
+      assert.equal(
+        serialized[k],
+        values[i],
+        `different serialized value ${k}, ${serialized[k]}, ${values[i]}`
+      )
+    }
+  })
+}
+
+module.exports = compareValues
